server/heavy/requirements.txt
appdirs==1.4.3
attrs==19.1.0
backcall==0.1.0
black==19.3b0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
decorator==4.4.0
defusedxml==0.6.0
Django==2.2.4
django-dotenv==1.4.2
django-extensions==2.2.1
django-filter==2.2.0
django-rest-framework==0.1.0
django-taggit==1.1.0
django-vote==2.1.7
djangorestframework==3.10.2
drf-extensions==0.5.0
idna==2.8
ipython==7.7.0
ipython-genutils==0.2.0
jedi==0.14.1
oauthlib==3.0.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.4.2
PyJWT==1.7.1
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
toml==0.10.0
traitlets==4.3.2
urllib3==1.25.3
wcwidth==0.1.7
